+++ lua/vibing/core/constants/init.lua
@@ -3,7 +3,6 @@
 local M = {}
 
 M.tools = require("vibing.core.constants.tools")
-M.actions = require("vibing.core.constants.actions")
 M.modes = require("vibing.core.constants.modes")
 
 return M